MDL-79316 Deleting Draft questions can break quiz
On the missing question placeholder using time() for the prepend part of questionid so it doesn't convert to a string.
Removed the exclusion of draft status from the question with references query to take them into account.
Additional check on the question capability to take into account missing questions.
Unit test